+++ applications/sample/mapwfs2/index.js
@@ -46,7 +46,7 @@ Oskari.clazz.define('Oskari.paikkatietoikkuna.Main', function() {
 
         var appConfig = this.appConfig;
         appConfig["mapfull"]["conf"]["plugins"][5]["id"] = "Oskari.mapframework.bundle.mapwfs2.plugin.WfsLayerPlugin";
-        appConfig["mapfull"]["conf"]["plugins"][5]["config"] = { contextPath : '/transport-0.0.1', port : '6060' }; // breaks if
+        appConfig["mapfull"]["conf"]["plugins"][5]["config"] = { contextPath : "/transport-0.0.1", hostname: "solmudmzke.nls.fi", port : "8888" };
         appConfig["mapfull"]["conf"]["plugins"][4]["config"] = { ignoredLayerTypes: ["WFS"], infoBox: false };
 
         appConfig["featuredata2"] = {conf: { selectionTools: true }};