+++ bundles/statistics/bundle/statsgrid/instance.js
@@ -42,6 +42,9 @@ function() {
         var request = sandbox.getRequestBuilder('userinterface.AddExtensionRequest')(this);
         sandbox.request(this, request);
 
+        var tooltipRequestHandler = Oskari.clazz.create('Oskari.statistics.bundle.statsgrid.request.TooltipContentRequestHandler', this);
+        sandbox.addRequestHandler('StatsGrid.TooltipContentRequest', tooltipRequestHandler);
+
         var locale = me.getLocalization();
         var mapModule = sandbox.findRegisteredModuleInstance('MainMapModule');
         this.mapModule = mapModule;
@@ -111,13 +114,7 @@ function() {
             }
         },
         'AfterMapLayerRemoveEvent': function(event) {
-            var layer = event.getMapLayer(),
-                layerId = layer.getId(),
-                view = this.plugins['Oskari.userinterface.View'];
-
-            if (layerId === view._layer.getId()) {
-                view.prepareMode(false);
-            }
+            this._afterMapLayerRemoveEvent(event);
         }
 	},
     /**
@@ -270,6 +267,24 @@ function() {
         this.state.colors = params.colors;
         // Send data to printout bundle
         this._createPrintParams(layer);
+    },
+
+    /**
+     * Exits the stats mode after the stats layer gets removed.
+     *
+     * @method _afterMapLayerRemoveEvent
+     * @private
+     * @param {Object} event
+     */
+    _afterMapLayerRemoveEvent: function(event) {
+        var layer = event.getMapLayer(),
+            layerId = layer.getId(),
+            view = this.plugins['Oskari.userinterface.View'];
+
+        // Exit the mode
+        if (view._layer && (layerId === view._layer.getId())) {
+            view.prepareMode(false);
+        }
     }
 }, {
 	"extend" : ["Oskari.userinterface.extension.DefaultExtension"]