+++ libraries/slickgrid/plugins/slick.checkboxselectcolumn.js
@@ -82,6 +82,7 @@
         }
 
         toggleRowSelection(args.row);
+
         e.stopPropagation();
         e.stopImmediatePropagation();
       }