+++ tests/statistics/statsgrid/statsgrid.js
@@ -12,7 +12,14 @@ describe('Test Suite for statistics/statsgrid bundle', function() {
 
     before(function() {
 
-        appSetup = getStartupSequence(['openlayers-default-theme', 'mapfull', 'divmanazer', 'toolbar', 'statsgrid']);
+        appSetup = getStartupSequence([
+            'openlayers-default-theme',
+            'mapfull',
+            'divmanazer',
+            'toolbar',
+            'statsgrid',
+            'printout'
+        ]);
 
         var mapfullConf = getConfigForMapfull();
         mapfullConf.conf.layers.push({
@@ -74,6 +81,7 @@ describe('Test Suite for statistics/statsgrid bundle', function() {
         it('should be defined', function() {
             expect(sandbox).to.be.ok();
             expect(statsModule).to.be.ok();
+            expect(viewPlugin).to.be.ok();
         });
     });
 
@@ -90,7 +98,6 @@ describe('Test Suite for statistics/statsgrid bundle', function() {
         after(teardown);
 
         it('should go to the mode view from the map view', function(done) {
-            // TODO: change spy to stub so that we test the interface and not the server
             var fetchCallbackGridSpy = sinon.spy(viewPlugin, 'createMunicipalitySlickGrid'),
                 fetchCallbackIndicatorSpy = sinon.spy(viewPlugin, 'createIndicatorsSelect');
 
@@ -130,24 +137,26 @@ describe('Test Suite for statistics/statsgrid bundle', function() {
                 return "Test.StatsGrid";
             }
             self.onEvent = function(event) {
-                var hasNaNs = false;
+                var hasNulls = false;
                 var params = event.getParams();
                 var colValues = params.COL_VALUES;
 
                 for (var i = 0; i < colValues.length; ++i) {
-                    if(isNaN(colValues[i])) {
-                        hasNaNs = true;
+                    if(colValues[i] == null) {
+                        hasNulls = true;
                     }
                 }
 
-                expect(hasNaNs).to.be(false);
+                expect(hasNulls).to.be(false);
 
                 // cleanup
                 sandbox.unregisterFromEventByName(self, 'StatsGrid.SotkadataChangedEvent');
+                sandbox._listeners = prevListeners;
                 done();
             }
 
             // Clear out other event listeners.
+            var prevListeners = sandbox._listeners;
             sandbox._listeners = {};
             // listen to StatsGrid.SotkadataChangedEvent to trigger verification
             sandbox.registerForEventByName(self, 'StatsGrid.SotkadataChangedEvent');
@@ -158,6 +167,7 @@ describe('Test Suite for statistics/statsgrid bundle', function() {
                     'fi': "Test indicator meta"
                 }
             });
+            // TODO: create a stub so that we test the interface and not the server
             gridPlugin.getSotkaIndicatorData(statsView.getEl(), 4, 'total', 2011);
         });
 
@@ -177,14 +187,76 @@ describe('Test Suite for statistics/statsgrid bundle', function() {
                 expect(eventData[0].url).to.be.ok();
 
                 sandbox.unregisterFromEventByName(self, 'Printout.PrintableContentEvent');
+                sandbox._listeners = prevListeners;
                 done();
             };
 
+            var prevListeners = sandbox._listeners;
             sandbox._listeners = {};
             sandbox.registerForEventByName(self, 'Printout.PrintableContentEvent');
 
             statsModule._createPrintParams(testLayer);
         });
+
+        it('should add checkbox column when statsgrid-show-row-selects checkbox is clicked', function(done) {
+            var fetchCallbackGridSpy = sinon.spy(viewPlugin, 'createMunicipalitySlickGrid'),
+                fetchCallbackIndicatorSpy = sinon.spy(viewPlugin, 'createIndicatorsSelect');
+
+            sandbox.postRequestByName('StatsGrid.StatsGridRequest', [true, testLayer]);
+            var i = 0;
+
+            waitsFor(function() {
+                return ((fetchCallbackGridSpy.callCount > 0) &&
+                        (fetchCallbackIndicatorSpy.callCount > 0));
+            }, function() {
+                menuToolbar = jQuery('body').find('div.oskariui-menutoolbar'),
+                statsGridContainer = jQuery('body').find('.statsgrid_100');
+
+                expect(statsGridContainer.is(':visible')).to.be(true);
+
+                //open drop down for statistical variable selector
+                jQuery('.slick-header-menubutton').click();
+                expect(jQuery('.slick-header-menu').css('visibility')).to.be('visible');
+
+                //open row selector checbox column
+                jQuery('.statsgrid-show-row-selects').find('input').click();
+                expect(jQuery('.slick-cell-checkboxsel').length).to.be.greaterThan(0);
+
+                //uncheck one of the municipalities and check if the name of the first group changes
+                expect(jQuery('.slick-group-title:first').text()).to.contain('Kunnat');
+                jQuery('.slick-cell-checkboxsel > input').first().prop('checked', false).click();
+                expect(jQuery('.slick-group-title:first').text()).to.contain('Poistettu');
+
+                fetchCallbackGridSpy.restore();
+                fetchCallbackIndicatorSpy.restore();
+                done();
+            }, "Waits for the stats grid mode request", 45000);
+        });
+
+
+
+        // OBS! This should be the last test case since we're removing the layer.
+        it('should exit the mode when a statistics layer gets removed', function(done) {
+            var statsView = statsModule.plugins['Oskari.userinterface.View'];
+            var removeLayerSpy = sinon.spy(statsModule, '_afterMapLayerRemoveEvent');
+            var exitModeSpy = sinon.spy(statsView, 'prepareMode');
+
+            // Remove the layer
+            var builder = sandbox.getRequestBuilder('RemoveMapLayerRequest');
+            var request = builder(testLayerId);
+            sandbox.request(statsModule, request);
+
+            setTimeout(function() {
+                expect(removeLayerSpy.callCount).to.be(1);
+                expect(exitModeSpy.callCount).to.be.greaterThan(0);
+                expect(exitModeSpy.calledWith(false)).to.be.ok();
+
+                removeLayerSpy.restore();
+                exitModeSpy.restore();
+                done();
+            }, 1000);
+        });
+
     });
 
 // TODO write test to: